+++ src/abstraction/alarm.cc
@@ -13,7 +13,7 @@ Alarm::Queue Alarm::_request;
 
 
 // Methods
-Alarm::Alarm(const Microsecond & time, Handler * handler, unsigned long times)
+Alarm::Alarm(const Microsecond & time, Handler * handler, int times)
 : _ticks(ticks(time)), _handler(handler), _times(times), _link(this, _ticks)
 {
     lock();
